Add files via upload
K nearest neighbors is a simple algorithm that stores all available cases and classifies new cases based on a similarity measure (e.g., distance functions). KNN has been used in statistical estimation and pattern recognition already in the beginning of 1970's as a non-parametric technique.

Algorithm
1.. Load the data
2.. Initialise the value of k
3.. For getting the predicted class, iterate from 1 to total number of training data points
    1.. Calculate the distance between test data and each row of training data. Here we will use Euclidean distance as our           distance metric since it’s the most popular method. The other metrics that can be used are Chebyshev, cosine, etc.
    2.. Sort the calculated distances in ascending order based on distance values
    3.. Get top k rows from the sorted array
    4.. Get the most frequent class of these rows
    5.. Return the predicted class
    
Choosing the number of nearest neighbors i.e. determining the value of k plays a significant role in determining the efficacy of the model. Thus, selection of k will determine how well the data can be utilized to generalize the results of the kNN algorithm. A large k value has benefits which include reducing the variance due to the noisy data; the side effect being developing a bias due to which the learner tends to ignore the smaller patterns which may have useful insights.