Add files via upload
In real-world data analysis tasks, we analyze complex data i.e. multidimensional data. We plot the data and find various patterns in it or use it to train some machine learning models. One way to think about dimensions is that suppose you have a data point x if we consider this data point as a physical object then dimensions are merely a basis of view, like where is the data located when it is observed from the horizontal axis or vertical axis.

As the dimensions of data increases, the difficulty to visualize it and perform computations on it also increases. So, how to reduce the dimensions of a data-
..Remove the redundant dimensions

..Only keep the most important dimensions

###what does PCA do?
PCA finds a new set of dimensions (or a set of the basis of views) such that all the dimensions are orthogonal (and hence linearly independent) and ranked according to the variance of data along with them. It means a more important principle
axis occurs first. (more important = more variance/more spread out data)

###How does PCA work -
1.. Calculate the covariance matrix X of data points.

2.. Calculate eigen vectors and corresponding eigen values.

3.. Sort the eigen vectors according to their eigen values in decreasing order.

4.. Choose first k eigen vectors and that will be the new k dimensions.

5.. Transform the original n dimensional data points into k dimensions.


###This goal of PCA -
1.. Find linearly independent dimensions (or basis of views) which can losslessly represent the data points.

2.. Those newly found dimensions should allow us to predict/reconstruct the original dimensions.The reconstruction/projection error should be minimized.